update sanitation, get four charts working
