+++ lib/sales_analyst.rb
@@ -1,15 +1,19 @@
 require 'date'
 require_relative 'statistics'
 require_relative 'item_analyst'
-require_relative 'merchant_analyst'
 require_relative 'invoice_analyst'
+require_relative 'merchant_item_analyst'
+require_relative 'merchant_invoice_analyst'
+require_relative 'merchant_revenue_analyst'
 require_relative 'make_charts'
 
 class SalesAnalyst
   include Statistics
   include ItemAnalyst
-  include MerchantAnalyst
   include InvoiceAnalyst
+  include MerchantItemAnalyst
+  include MerchantInvoiceAnalyst
+  include MerchantRevenueAnalyst
   include MakeCharts
 
   attr_reader :engine,
